Add files via upload
Big data utilization has become a major component of the crime analysis process. Different sectors have been adopting big data technologies which has led to an increase in efficiency and effectiveness and the crime sector is no different. Big data technology has been adopted in crime analysis, from crime patterns and predictions and automating the fraud detection process.
This report adopts a CRISP_DM methodology to answer the research questions and also a systematic review methodology to examine the challenges and solutions associated with the introduction of big data technologies.
The document examines the datasets, highlights the issues associated with working with the datasets, and provides solutions adopted to handle the issues. The report emphasizes the importance of big data technologies in crime analysis bringing us to a conclusion that there are more benefits to the introduction of big data solution.